+++ otlet/__init__.py
@@ -27,7 +27,7 @@ CLI tool and wrapper for the PyPI JSON web API
 from .api import *
 from .exceptions import *
 
-__version__ = "1.0.0rc1"
+__version__ = "1.0.0rc2"
 __license__ = "MIT"
 
 DEPENDENCY_ENVIRONMENT_MARKERS = [