+++ otlet/packaging/version/__init__.py
@@ -311,6 +311,13 @@ class Version(_BaseVersion):
             parts.append(f"+{self.local}")
 
         return "".join(parts)
+    
+    def fits_constraints(self, constraints: list) -> bool:
+        _c = []
+        for i in constraints:
+            _h = re.match(r'([=<>!]+)(\S+)', i)
+            exec(f"_c.append(self {_h.group(1)} self.__class__(\"{_h.group(2)}\"))")
+        return not any([not _ for _ in _c])
 
     @property
     def epoch(self) -> int: